requirements.txt
aiofiles==0.6.0
certifi==2020.11.8
click==7.1.2
deta==0.7
fastapi==0.61.2
h11==0.11.0
httpcore==0.12.1
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
pydantic==1.7.2
python-dateutil==2.8.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.12.2
